requirements.txt
# 웹 서버를 만들기 위한 프레임워크 (Cochl Webhook을 받기 위해 필요)
fastapi==0.109.0

# FastAPI를 실행하기 위한 서버
uvicorn[standard]==0.27.0

# HTTP 요청을 보내기 위한 라이브러리 (Zapier Webhook 호출용)
requests==2.31.0

# 환경 변수 관리를 위한 라이브러리 (API 키를 안전하게 관리)
python-dotenv==1.0.0

# 데이터 검증을 위한 라이브러리 (FastAPI에서 기본 사용)
pydantic==2.5.3

# 날짜/시간 처리를 위한 라이브러리
python-dateutil==2.8.2

# 파일 업로드 지원을 위한 라이브러리
python-multipart==0.0.6

# 비동기 HTTP 클라이언트 (Cochl API 호출용)
httpx==0.26.0

# 비동기 파일 처리
aiofiles==23.2.1
